--- source/source/checkprerequisits.h
@@ -1,8 +1,0 @@
-#ifndef __CHECKBASICS_H
-#define __CHECKBASICS_H
-
-
-void check_prerequisits();
-
-
-#endif
