+++ source/source/drunner_setup.cpp
@@ -7,7 +7,6 @@
 #include "globallogger.h"
 #include "globalcontext.h"
 #include "drunner_paths.h"
-#include "checkprerequisits.h"
 #include "validateimage.h"
 
 #include <Poco/Process.h>
@@ -16,6 +15,71 @@
 
 namespace drunnerSetup
 {
+
+   void _check_prereqs_xplatform()
+   {
+      CommandLine cl("docker", { "--version" });
+      std::string op;
+      if (utils::runcommand(cl, op, utils::kRC_Defaults) != 0)
+         fatal("Running \"docker --version\" failed!\nIs docker correctly installed on this machine?\n" + op);
+   }
+
+   // ---------------------------------------------------------------------------------------------------------------------------------------------
+   // ---------------------------------------------------------------------------------------------------------------------------------------------
+   // ---------------------------------------------------------------------------------------------------------------------------------------------
+
+
+#ifdef _WIN32
+   void _check_prerequisits() 
+   { 
+      _check_prereqs_xplatform(); 
+   }
+#else   // Linux
+#include <unistd.h>
+
+   int bashcommand(std::string bashline, std::string & op, bool trim)
+   {
+      CommandLine cl("/bin/bash", { "-c", bashline });
+      return utils::runcommand(cl, op, utils::kRC_Trim);
+   }
+   int bashcommand(std::string bashline)
+   {
+      std::string op;
+      return bashcommand(bashline, op, false);
+   }
+   bool commandexists(std::string command)
+   {
+      return (0 == bashcommand("command -v " + command));
+   }
+   std::string getUSER()
+   {
+      char s[200];
+      getlogin_r(s, 199);
+      return s;
+   }
+
+   void _check_prerequisits()
+   {
+      uid_t euid = geteuid();
+      if (euid == 0)
+         fatal("Please run as a standard user, not as root.");
+
+      std::string user = getUSER();
+      if (0 != bashcommand("groups $USER | grep docker"))
+         fatal("Please add the current user to the docker group. As root: " + utils::kCODE_S + "adduser " + user + " docker" + utils::kCODE_E);
+
+      if (0 != bashcommand("groups | grep docker"))
+         fatal(user + " hasn't picked up group docker yet. Log out then in again, or run " + utils::kCODE_S + "exec su -l " + user + utils::kCODE_E);
+
+      if (!commandexists("docker"))
+         fatal("Please install Docker before using dRunner.\n(e.g. use  https://raw.githubusercontent.com/j842/scripts/master/install_docker.sh )");
+
+      if (!commandexists("curl"))
+         fatal("Please install curl before using dRunner.");
+
+      _check_prereqs_xplatform();
+   }
+#endif
 
    void _copyexe()
    {
@@ -56,7 +120,7 @@ namespace drunnerSetup
       s_setup_fully_checked = true;
 
       // check prereqs (e.g. docker installed).
-      check_prerequisits();
+      _check_prerequisits();
 
       // -----------------------------------------------------------------------------
       // create directory structure.
@@ -91,12 +155,4 @@ namespace drunnerSetup
       // Finished!
       return kRSuccess;
    }
-
-   cResult setup()
-   {
-      cResult rval = check_setup(true);
-      logmsg(kLINFO, "Setup complete.");
-      return rval;
-   }
-
 }
