+++ vs/drunner.vcxproj
@@ -54,7 +54,6 @@
     <ClCompile Include="..\source\deps\lua\lzio.c" />
     <ClCompile Include="..\source\service_lua_cfuncs.cpp" />
     <ClCompile Include="..\source\source\buildnum.cpp" />
-    <ClCompile Include="..\source\source\checkprerequisits.cpp" />
     <ClCompile Include="..\source\source\command_dev.cpp" />
     <ClCompile Include="..\source\source\command_general.cpp" />
     <ClCompile Include="..\source\source\drunner_paths.cpp" />
@@ -120,7 +119,6 @@
     <ClInclude Include="..\source\deps\lua\lzio.h" />
     <ClInclude Include="..\source\source\basen.h" />
     <ClInclude Include="..\source\source\buildnum.h" />
-    <ClInclude Include="..\source\source\checkprerequisits.h" />
     <ClInclude Include="..\source\source\chmod.h" />
     <ClInclude Include="..\source\source\command_dev.h" />
     <ClInclude Include="..\source\source\command_general.h" />
