+++ vs/drunner.vcxproj.filters
@@ -88,9 +88,6 @@
     <ClCompile Include="..\source\source\drunner_setup.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="..\source\source\checkprerequisits.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
     <ClCompile Include="..\source\source\validateimage.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
@@ -310,9 +307,6 @@
       <Filter>Source Files</Filter>
     </ClInclude>
     <ClInclude Include="..\source\source\drunner_setup.h">
-      <Filter>Source Files</Filter>
-    </ClInclude>
-    <ClInclude Include="..\source\source\checkprerequisits.h">
       <Filter>Source Files</Filter>
     </ClInclude>
     <ClInclude Include="..\source\source\validateimage.h">
